Allow ourselves to handle String objects bigger than stack size
My default stack for a Ruby process is around 8M (Debian Etch on
x86), so I'm not sure how sane this is (manipulating large
Strings in Ruby is insane), but some people have have much
smaller stack sizes (or they are insane).

So we'll estimate how much we can allocate onto the stack based
on the stack size rlimit gives us, and use a malloc(3)/free(3)
pair for memory allocations that can't fit on the stack.

I'd use mmap(2) with a file as a backing store, but not all Ruby
platforms support that (of course people running those insane
platforms are probably the ones dealing with insanely large
String objects, too :)

Enable the tests for these by setting LARGE_STRING_TEST in the
environment.